@@ -1,7 +1,9 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <WarningsAsErrors>nullable</WarningsAsErrors>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
@@ -16,9 +18,7 @@
     <None Remove="broker.db" />
     <None Remove="broker.db-shm" />
     <None Remove="broker.db-wal" />
-    <None Remove="cert.pem" />
     <None Remove="fullchain1.pem" />
-    <None Remove="key.pem" />
   </ItemGroup>
 
   <ItemGroup>
@@ -26,21 +26,32 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="H.Socket.IO" Version="1.2.1" />
     <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
     <PackageReference Include="Microsoft.AspNetCore.SignalR.Protocols.Json" Version="3.1.0-preview3.19555.2" />
     <PackageReference Include="Microsoft.AspNetCore.SignalR.Protocols.NewtonsoftJson" Version="3.1.0-preview3.19555.2" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.0-preview3.19554.8" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="3.1.0-preview3.19554.8" />
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
     <PackageReference Include="NLog" Version="4.6.8" />
-    <PackageReference Include="SimpleSocketIoClient" Version="0.9.4.2" />
     <PackageReference Include="System.Text.Encoding.CodePages" Version="4.7.0-preview3.19551.4" />
     <PackageReference Include="System.Text.Encodings.Web" Version="4.7.0-preview3.19551.4" />
-    <PackageReference Include="System.Text.Json" Version="4.7.0-preview3.19551.4" />
+    <PackageReference Include="Azura" Version="0.7.2" />
+    <PackageReference Include="Azura.Generator" Version="0.7.2" />
+
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\PainlessMesh\PainlessMesh.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="cert.pem">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="key.pem">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 
 </Project>
