@@ -21,70 +21,84 @@ using AppBokerASP.Devices.Zigbee;
 using System.Buffers.Text;
 using NLog;
 
-namespace AppBokerASP.Devices.Heater
+namespace AppBokerASP.Devices.Painless.Heater
 {
-    public class Heater : Device, IDisposable
+    [PainlessMeshName("heater")]
+
+    public class Heater : PainlessDevice, IDisposable
     {
 
-        public HeaterConfig Temperature { get; set; }
-        public long XiaomiTempSensor { get; set; }
-        public HeaterConfig CurrentConfig { get; set; }
-        public HeaterConfig CurrentCalibration { get; private set; }
-        public string FirmwareVersion { get; private set; }
+        public HeaterConfig? Temperature { get; set; }
+        public long XiaomiTempSensor { get; set; } = 0;
+        public HeaterConfig? CurrentConfig { get; set; }
+        public HeaterConfig? CurrentCalibration { get; private set; }
 
-        private readonly List<HeaterConfig> timeTemps = new List<HeaterConfig>();
+        private readonly List<HeaterConfig> timeTemps = new();
 
-        private string logName => Id + "/" + FriendlyName;
 
-        private Task heaterSensorMapping;
+        private readonly Task? heaterSensorMapping;
 
         //2020-02-01 20:14:48.1075|DEBUG|AppBokerASP.BaseClient|{"id":3257233774, "m":"Update", "c":"WhoIAm", "p":["10.12.206.9","heater","RmlybXdhcmUgVjIgRmViICAxIDIwMjA=","YAk3oJQqQBo3QKkqYQk3oZQqQRo3QakqYgk3opQqQho3QqkqYwk3o5QqQxo3Q6kqZAk3pJQqRBo3RKkqJQ03RakqJg03Rqkq"]}
-        public Heater(long id, List<string> parameters) : base(id)
+        public Heater(long id, ByteLengthList parameters) : base(id)
         {
-            Program.MeshManager.SingleUpdateMessageReceived += Node_SingleUpdateMessageReceived;
-            Program.MeshManager.SingleOptionsMessageReceived += Node_SingleOptionsMessageReceived;
             using var cont = DbProvider.BrokerDbContext;
-            timeTemps.AddRange(cont.HeaterConfigs.Where(x => x.Device.Id == id).ToList().Select(x => (HeaterConfig)x));
-            InterpretParameters(parameters);
+            timeTemps.AddRange(cont.HeaterConfigs.Where(x => x.Device!.Id == id).ToList().Select(x => (HeaterConfig)x));
 
             ShowInApp = true;
             //cont.HeaterCalibrations.FirstOrDefault(x => x.Id == Id);
+            try
+            {
 
-            var heater = cont.Devices.FirstOrDefault(x => x.Id == Id);
-            var mappings = cont.DeviceToDeviceMappings.Include(x => x.Child).Where(x => x.Parent.Id == id /*&& cont.Devices.Any(y => y.Id == x.Child.Id)*/).ToList();
-            logger.Debug($"Heater {logName} has {mappings.Count} mappings");
-            if (mappings.Count > 0)
-                heaterSensorMapping = Task.Run(() => TrySubscribe(mappings));
+                var heater = cont.Devices.FirstOrDefault(x => x.Id == Id);
+                var mappings = cont.DeviceToDeviceMappings.Include(x => x.Child).Where(x => x.Parent!.Id == id /*&& cont.Devices.Any(y => y.Id == x.Child.Id)*/).ToList();
+                logger.Debug($"Heater {LogName} has {mappings?.Count} mappings");
+                if (mappings?.Count > 0)
+                    heaterSensorMapping = Task.Run(() => TrySubscribe(mappings));
+            }
+            catch (Exception ex)
+            {
+
+            }
         }
 
-        private void InterpretParameters(List<string> parameters)
+        protected override void InterpretParameters(ByteLengthList parameters)
         {
-            if (parameters.Count > 2)
-            {
-                FirmwareVersion = Encoding.ASCII.GetString(Convert.FromBase64String(parameters[2]));
-            }
+            if (parameters is null)
+                return;
+            base.InterpretParameters(parameters);
+            //if (parameters.Count > 2)
+            //{
+            //var s2 = "";
+            //timeTemps.OrderBy(x => x.DayOfWeek).ThenBy(x => x.TimeOfDay).ToList().ForEach(x => s2 += ((TimeTempMessageLE)x).ToString());
+            //var msg = new GeneralSmarthomeMessage((uint)Id, MessageType.Options, Command.Temp, s2.ToJToken());
+            //Program.MeshManager.SendSingle((uint)Id, msg);
+
+            //}
             if (parameters.Count > 3)
             {
                 try
                 {
                     var s = parameters[3];
-                    var s2 = "";
-                    timeTemps.OrderBy(x=>x.DayOfWeek).ThenBy(x=>x.TimeOfDay).ToList().ForEach(x => s2 += ((TimeTempMessageLE)x).ToString());
+                    byte[] s2 = GetSendableTimeTemps(timeTemps);
 
-                    if (s != s2)
+                    if (!s.SequenceEqual(s2))
                     {
-                        logger.Warn($"Heater {logName} has wrong temps saved, trying to correct Saved:{{{s}}} Server:{{{s2}}}");
-                        var msg = new GeneralSmarthomeMessage((uint)Id, MessageType.Options, Command.Temp, s2.ToJToken());
+                        logger.Warn($"Heater {LogName} has wrong temps saved, trying correcting");
+                        var msg = new BinarySmarthomeMessage((uint)Id, MessageType.Options, Command.Temp, s2);
                         Program.MeshManager.SendSingle((uint)Id, msg);
                     }
                 }
                 catch (Exception e)
                 {
-                    logger.Error(e, $"Heater {logName} has transmitted wrong temps");
+                    logger.Error(e, $"Heater {LogName} has transmitted wrong temps");
                 }
             }
         }
 
+        private byte[] GetSendableTimeTemps(IEnumerable<HeaterConfig> timeTemps)
+        {
+            return timeTemps.OrderBy(x => x.DayOfWeek).ThenBy(x => x.TimeOfDay).SelectMany(x => ((TimeTempMessageLE)x).ToBinary()).ToArray();
+        }
 
         private Task TrySubscribe(List<DeviceMappingModel> mappings)
         {
@@ -94,7 +108,7 @@ namespace AppBokerASP.Devices.Heater
                 foreach (var mapping in mappings)
                 {
 
-                    if (mapping?.Child == null)
+                    if (mapping is not null && mapping.Child is null)
                     {
                         toRemove.Add(mapping);
                         break;
@@ -106,18 +120,20 @@ namespace AppBokerASP.Devices.Heater
                         if (device is XiaomiTempSensor sensor)
                         {
                             sensor.TemperatureChanged += XiaomiTempSensorTemperaturChanged;
+                            SendLastTempData();
                             XiaomiTempSensor = device.Id;
                         }
-                        logger.Debug($"Heater {logName} has subscribed to {device.Id}/{device.FriendlyName}");
+                        logger.Debug($"Heater {LogName} has subscribed to {device.Id}/{device.FriendlyName}");
 
-                        toRemove.Add(mapping);
+                        if (mapping is not null)
+                            toRemove.Add(mapping);
                         break;
                     }
                 }
 
                 foreach (var mapping in toRemove)
                 {
-                    mappings.Remove(mapping);
+                    _ = mappings.Remove(mapping);
                 }
 
                 Thread.Sleep(1000);
@@ -125,26 +141,21 @@ namespace AppBokerASP.Devices.Heater
             return Task.CompletedTask;
         }
 
-        private void Node_SingleUpdateMessageReceived(object sender, GeneralSmarthomeMessage e)
+        protected override void UpdateMessageReceived(BinarySmarthomeMessage e)
         {
-            if (e.NodeId != Id)
-                return;
-            logger.Debug($"DataReceived in {nameof(Node_SingleUpdateMessageReceived)} {logName}: " + e.ToJson());
+
             switch (e.Command)
             {
                 case Command.Temp:
-                    HandleTimeTempMessageUpdate((string)e.Parameters[0]);
+                    HandleTimeTempMessageUpdate(e.Parameters[0]);
                     break;
                 default:
                     break;
             }
         }
-        private void Node_SingleOptionsMessageReceived(object sender, GeneralSmarthomeMessage e)
-        {
-            if (e.NodeId != Id)
-                return;
 
-            logger.Debug($"DataReceived in {nameof(Node_SingleOptionsMessageReceived)} {logName}: " + e.ToJson());
+        protected override void OptionMessageReceived(BinarySmarthomeMessage e)
+        {
             switch (e.Command)
             {
                 case Command.Temp:
@@ -156,24 +167,22 @@ namespace AppBokerASP.Devices.Heater
             }
         }
 
-        private void HandleTimeTempMessageUpdate(string messages)
+        private void HandleTimeTempMessageUpdate(byte[] messages)
         {
-            var ttm = TimeTempMessageLE.FromBase64(messages[0..4]);
+            var message = messages.AsSpan();
+            var ttm = TimeTempMessageLE.LoadFromBinary(message[0..3]);
             Temperature = ttm;
-            ttm = TimeTempMessageLE.FromBase64(messages[4..8]);
+            ttm = TimeTempMessageLE.LoadFromBinary(message[3..6]);
             CurrentConfig = ttm;
-            ttm = TimeTempMessageLE.FromBase64(messages[8..12]);
+            ttm = TimeTempMessageLE.LoadFromBinary(message[6..9]);
             try
             {
-                var dt = DateTime.Now;
-                dt.AddHours(ttm.Time.Hours - dt.Hour);
-                dt.AddMinutes(ttm.Time.Minutes - dt.Minute);
-
-                CurrentCalibration = new HeaterConfig(ttm.DayOfWeek, dt, ttm.Temp - 51.2f);
+                CurrentCalibration = ttm;
+                CurrentCalibration.Temperature -= 51.2f;
             }
             catch (Exception e)
             {
-                logger.Error($"Heater {logName} has Exception inside HandleTimeTempMessageUpdate\r\n {e} ");
+                logger.Error($"Heater {LogName} has Exception inside HandleTimeTempMessageUpdate\r\n {e} ");
                 Console.WriteLine(e);
             }
 
@@ -190,7 +199,8 @@ namespace AppBokerASP.Devices.Heater
                 case Command.Temp:
                     var temp = (float)parameters[0];
                     var ttm = new TimeTempMessageLE((DayOfWeek)((((byte)DateTime.Now.DayOfWeek) + 6) % 7), new TimeSpan(DateTime.Now.TimeOfDay.Hours, DateTime.Now.TimeOfDay.Minutes, 0), temp);
-                    var msg = new GeneralSmarthomeMessage((uint)Id, MessageType.Update, command, ttm.ToString().ToJToken());
+                    //logger.Debug("Send new ttm: " + )
+                    var msg = new BinarySmarthomeMessage((uint)Id, MessageType.Update, command, ttm.ToBinary());
                     Program.MeshManager.SendSingle((uint)Id, msg);
                     break;
                 case Command.DeviceMapping:
@@ -206,27 +216,26 @@ namespace AppBokerASP.Devices.Heater
         {
 
             logger.Debug("OptionsFromApp " + command + " <> " + parameters.ToJson());
-            GeneralSmarthomeMessage msg;
+            BinarySmarthomeMessage msg;
+
             switch (command)
             {
                 case Command.Temp:
-                {
-                    var hc = parameters.Select(x => x.ToDeObject<HeaterConfig>()).OrderBy(x => x.DayOfWeek).ThenBy(x => x.TimeOfDay);
+                    {
+                        var hc = parameters.Select(x => x.ToDeObject<HeaterConfig>()).OrderBy(x => x.DayOfWeek).ThenBy(x => x.TimeOfDay);
 
-                    UpdateDB(hc);
+                        UpdateDB(hc);
 
-                    var ttm = hc.Select(x => new TimeTempMessageLE(x.DayOfWeek, new TimeSpan(x.TimeOfDay.Hour, x.TimeOfDay.Minute, 0), (float)x.Temperature));
-                    var s = "";
-                    foreach (var item in ttm)
-                        s += item.ToString();
+                        var ttm = hc.Select(x => new TimeTempMessageLE(x.DayOfWeek, new TimeSpan(x.TimeOfDay.Hour, x.TimeOfDay.Minute, 0), (float)x.Temperature));
+                        var s = ttm.SelectMany(x => x.ToBinary()).ToArray();
 
-                    msg = new GeneralSmarthomeMessage((uint)Id, MessageType.Options, command, s.ToJToken());
-                    Program.MeshManager.SendSingle((uint)Id, msg);
-                    break;
-                }
+                        msg = new BinarySmarthomeMessage((uint)Id, MessageType.Options, command, s);
+                        Program.MeshManager.SendSingle((uint)Id, msg);
+                        break;
+                    }
 
                 case Command.Mode:
-                    msg = new GeneralSmarthomeMessage((uint)Id, MessageType.Options, command);
+                    msg = new BinarySmarthomeMessage((uint)Id, MessageType.Options, command);
                     Program.MeshManager.SendSingle((uint)Id, msg);
                     break;
             }
@@ -240,17 +249,18 @@ namespace AppBokerASP.Devices.Heater
             var models = hc.Select(x => (HeaterConfigModel)x).ToList();
             using var cont = DbProvider.BrokerDbContext;
             var d = cont.Devices.FirstOrDefault(x => x.Id == Id);
-            foreach (var item in models)
-                item.Device = d;
+            if (d is not null)
+                foreach (var item in models)
+                    item.Device = d;
 
-            var oldConfs = cont.HeaterConfigs.Where(x => x.Device.Id == Id).ToList();
+            var oldConfs = cont.HeaterConfigs.Where(x => x.Device!.Id == Id).ToList();
             if (oldConfs.Count > 0)
             {
                 cont.RemoveRange(oldConfs);
-                cont.SaveChanges();
+                _ = cont.SaveChanges();
             }
             cont.AddRange(models);
-            cont.SaveChanges();
+            _ = cont.SaveChanges();
         }
 
         private void UpdateDeviceMappingInDb(long tempId, long oldId)
@@ -262,32 +272,33 @@ namespace AppBokerASP.Devices.Heater
             if (heater == default || tempSensor == default || sensor == default)
                 return;
 
-
-            var oldMappings = cont.DeviceToDeviceMappings.Where(x => x.Parent.Id == Id);
+            var oldMappings = cont.DeviceToDeviceMappings.Where(x => x.Parent!.Id == Id);
             foreach (var oldMapping in oldMappings)
             {
                 var oldsensor = Program.DeviceManager.Devices.FirstOrDefault(x => x.Key == oldId).Value;
-                if (oldsensor != default)
-                    ((XiaomiTempSensor)oldsensor).TemperatureChanged -= XiaomiTempSensorTemperaturChanged;
-                cont.Remove(oldMapping);
+                if (oldsensor is not null and XiaomiTempSensor xts)
+                    xts.TemperatureChanged -= XiaomiTempSensorTemperaturChanged;
+                _ = cont.Remove(oldMapping);
             }
-            cont.SaveChanges();
+            _ = cont.SaveChanges();
 
 
-            cont.Add(new DeviceMappingModel
+            _ = cont.Add(new DeviceMappingModel
             {
                 Parent = heater,
                 Child = tempSensor
             });
-            (sensor as XiaomiTempSensor).TemperatureChanged += XiaomiTempSensorTemperaturChanged;
+
+            if (sensor is XiaomiTempSensor xiaomiTempSensor)
+                xiaomiTempSensor.TemperatureChanged += XiaomiTempSensorTemperaturChanged;
             XiaomiTempSensor = sensor.Id;
-            cont.SaveChanges();
+            _ = cont.SaveChanges();
         }
 
-        private void XiaomiTempSensorTemperaturChanged(object sender, float temp)
+        private void XiaomiTempSensorTemperaturChanged(object? sender, float temp)
         {
             var ttm = new TimeTempMessageLE((DayOfWeek)((((byte)DateTime.Now.DayOfWeek) + 6) % 7), new TimeSpan(DateTime.Now.TimeOfDay.Hours, DateTime.Now.TimeOfDay.Minutes, 0), temp);
-            var msg = new GeneralSmarthomeMessage((uint)Id, MessageType.Relay, Command.Temp, ttm.ToString().ToJToken());
+            var msg = new BinarySmarthomeMessage((uint)Id, MessageType.Relay, Command.Temp, ttm.ToBinary());
             Program.MeshManager.SendSingle((uint)Id, msg);
         }
 
@@ -296,24 +307,34 @@ namespace AppBokerASP.Devices.Heater
 
         }
 
-        public override void StopDevice()
+
+
+        public override void Reconnect(ByteLengthList parameter)
         {
-            base.StopDevice();
+            base.Reconnect(parameter);
+
+            InterpretParameters(parameter);
+
             Program.MeshManager.SingleUpdateMessageReceived -= Node_SingleUpdateMessageReceived;
             Program.MeshManager.SingleOptionsMessageReceived -= Node_SingleOptionsMessageReceived;
-        }
-
-        public override void Reconnect()
-        {
-            base.Reconnect();
             Program.MeshManager.SingleUpdateMessageReceived += Node_SingleUpdateMessageReceived;
             Program.MeshManager.SingleOptionsMessageReceived += Node_SingleOptionsMessageReceived;
+
+            SendLastTempData();
         }
+
 
         public override dynamic GetConfig() => timeTemps.ToJson();
         public void Dispose()
         {
             heaterSensorMapping?.Dispose();
+        }
+
+        private void SendLastTempData()
+        {
+            if (Program.DeviceManager.Devices.TryGetValue(XiaomiTempSensor, out var device)
+                    && device is XiaomiTempSensor sensor && sensor.Temperature > 5f)
+                XiaomiTempSensorTemperaturChanged(this, sensor.Temperature);
         }
     }
 }
