@@ -1,73 +1,133 @@
 ﻿using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Net;
 using System.Net.NetworkInformation;
 using System.Net.Sockets;
-using System.Text.Json;
+using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+
+using AppBokerASP.Devices.Painless;
+
+using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+
 using PainlessMesh;
 
 namespace AppBokerASP
 {
 
-    public class SmarthomeMeshManager
+    public class SmarthomeMeshManager : IDisposable
     {
         private class NodeSync
         {
-            public long Id { get; }
+            public uint Id { get; }
             public byte MissedConnections { get; set; }
-            public NodeSync(long id, byte missedConnections)
+            public NodeSync(uint id, byte missedConnections)
             {
                 Id = id;
                 MissedConnections = missedConnections;
             }
         }
 
-        public event EventHandler<GeneralSmarthomeMessage> SingleUpdateMessageReceived;
-        public event EventHandler<GeneralSmarthomeMessage> SingleOptionsMessageReceived;
-        public event EventHandler<GeneralSmarthomeMessage> SingleGetMessageReceived;
-        public event EventHandler<(Sub, List<string>)> NewConnectionEstablished;
-        public event EventHandler<long> ConnectionLost;
-        public event EventHandler<long> ConnectionReastablished;
+        public event EventHandler<BinarySmarthomeMessage>? SingleUpdateMessageReceived;
+        public event EventHandler<BinarySmarthomeMessage>? SingleOptionsMessageReceived;
+        public event EventHandler<BinarySmarthomeMessage>? SingleGetMessageReceived;
+        public event EventHandler<(Sub, ByteLengthList)>? NewConnectionEstablished;
+        public event EventHandler<uint>? ConnectionLost;
+        public event EventHandler<(uint id, ByteLengthList parameter)>? ConnectionReastablished;
 
-        private static ServerSocket serverSocket = new ServerSocket();
+        private static readonly ServerSocket serverSocket = new();
 
         private readonly TimeSpan WaitBeforeWhoIAmSendAgain;
         private readonly List<NodeSync> knownNodeIds;
-        private readonly ConcurrentDictionary<long, (DateTime time, int count)> WhoIAmSendTime;
+        private readonly int listenPort;
+        private readonly ConcurrentDictionary<uint, (DateTime time, int count)> WhoIAmSendTime;
 
-        private readonly long nodeID = 1;
-        private readonly Dictionary<long, Queue<GeneralSmarthomeMessage>> queuedMessages;
+        private readonly uint nodeID = 1;
+        private readonly Dictionary<uint, Queue<BinarySmarthomeMessage>> queuedMessages;
         private readonly List<Timer> timers;
         private readonly NLog.Logger logger = NLog.LogManager.GetCurrentClassLogger();
+        private bool running;
 
-
-        public SmarthomeMeshManager(int listenPort, long nodeId = 1)
+        public SmarthomeMeshManager(int listenPort, uint nodeId = 1)
         {
-            WaitBeforeWhoIAmSendAgain = new TimeSpan(0, 0, 30);
-            queuedMessages = new Dictionary<long, Queue<GeneralSmarthomeMessage>>();
-            timers = new List<Timer>();
+            WaitBeforeWhoIAmSendAgain = new TimeSpan(0, 0, 10);
+
+            queuedMessages = new();
+            timers = new();
             nodeID = nodeId;
-            WhoIAmSendTime = new ConcurrentDictionary<long, (DateTime, int)>();
-            knownNodeIds = new List<NodeSync> { new NodeSync(nodeID, 0), new NodeSync(0, 0) };
+            WhoIAmSendTime = new();
+            knownNodeIds = new() { new NodeSync(nodeID, 0), new NodeSync(0, 0) };
+            this.listenPort = listenPort;
+
+        }
+
+        public void Start()
+        {
+            if (running)
+                return;
+            running = true;
             serverSocket.OnClientConnected += ServerSocket_OnClientConnected;
             serverSocket.Start(new IPAddress(new byte[] { 0, 0, 0, 0 }), listenPort);
-            timers.Add(new Timer(WhoAmITask, null, TimeSpan.FromSeconds(20d), WaitBeforeWhoIAmSendAgain));
-            timers.Add(new Timer((n) => SendBroadcast(new GeneralSmarthomeMessage(0, MessageType.Update, Command.Time, $"{{\"Date\":\"{DateTime.Now:dd.MM.yyyy HH:mm:ss}\"}}".ToJToken())), null, TimeSpan.FromMinutes(1d), TimeSpan.FromHours(1d)));
-            timers.Add(new Timer((n) => SendToBridge(new GeneralSmarthomeMessage(1, MessageType.Get, Command.Mesh)), null, TimeSpan.FromSeconds(10d), TimeSpan.FromMinutes(1d)));
+
+            timers.Add(new Timer(TryCatchWhoAmITask, null, TimeSpan.FromSeconds(10d), WaitBeforeWhoIAmSendAgain));
+            timers.Add(new Timer(SendTimeUpdate, null, TimeSpan.FromMinutes(1d), TimeSpan.FromHours(1d)));
+            timers.Add(new Timer(GetMeshUpdate, null, TimeSpan.FromSeconds(10d), TimeSpan.FromSeconds(10d)));
+            Program.UpdateManager.Advertisment += OtaAdvertisment;
         }
 
-        private void ServerSocket_OnClientConnected(object sender, BaseClient baseClient)
+        public void Stop()
+        {
+            if (!running)
+                return;
+            running = false;
+            serverSocket.Stop();
+            serverSocket.OnClientConnected -= ServerSocket_OnClientConnected;
+
+            foreach (var timer in timers)
+            {
+                timer.Dispose();
+            }
+            timers.Clear();
+        }
+
+        private void GetMeshUpdate(object? state)
+        {
+            SendToBridge(new BinarySmarthomeMessage(1, MessageType.Get, Command.Mesh));
+        }
+
+        internal void UpdateTime() => SendTimeUpdate(null);
+
+        private static readonly byte[] fromServer = new byte[] { 1 };
+        private void SendTimeUpdate(object? state)
+        {
+            var dto = new DateTimeOffset(DateTime.Now.Ticks, TimeSpan.Zero);
+            var list = new ByteLengthList(fromServer, BitConverter.GetBytes((int)dto.ToUnixTimeSeconds()));
+            var msg = new BinarySmarthomeMessage(0, MessageType.Update, Command.Time, list);
+            SendBroadcast(msg);
+        }
+
+
+        private void OtaAdvertisment(object? sender, PainlessMesh.Ota.FirmwareMetadata e)
+        {
+            foreach (var item in Program.DeviceManager.Devices.Values)
+            {
+                if (item is PainlessDevice pd)
+                    pd.OtaAdvertisment(e);
+            }
+        }
+
+        private void ServerSocket_OnClientConnected(object? sender, BaseClient baseClient)
         {
             baseClient.ReceivedData += SocketClientDataReceived;
-            SendToBridge(new GeneralSmarthomeMessage(1, MessageType.Get, Command.Mesh));
+            SendToBridge(new BinarySmarthomeMessage(1, MessageType.Get, Command.Mesh));
         }
 
-        public void SocketClientDataReceived(object sender, GeneralSmarthomeMessage e)
+        public void SocketClientDataReceived(object? sender, BinarySmarthomeMessage e)
         {
             //var bc = (BaseClient)sender;
 
@@ -76,11 +136,12 @@ namespace AppBokerASP
 
             if (e.Command == Command.WhoIAm)
             {
-                knownNodeIds.Add(new NodeSync(e.NodeId, 0));
-                while (WhoIAmSendTime.ContainsKey(e.NodeId) && !WhoIAmSendTime.TryRemove(e.NodeId, out var asda)) { }
-                if (e.Parameters == null)
-                    return;
-                NewConnectionEstablished?.Invoke(this, (new Sub { NodeId = e.NodeId }, e.Parameters?.ToStringArray().ToList()));
+                if (!knownNodeIds.Any(x => x.Id == e.NodeId))
+                    knownNodeIds.Add(new NodeSync(e.NodeId, 0));
+
+                _ = WhoIAmSendTime.TryRemove(e.NodeId, out var asda);
+                if (e.Parameters != null)
+                    NewConnectionEstablished?.Invoke(this, (new Sub { NodeId = e.NodeId }, e.Parameters));
                 return;
             }
 
@@ -89,15 +150,15 @@ namespace AppBokerASP
             {
                 if (!WhoIAmSendTime.TryGetValue(e.NodeId, out var dt) || dt.time.Add(WaitBeforeWhoIAmSendAgain) > DateTime.Now)
                 {
-                    //SendSingle(e.NodeId, new GeneralSmarthomeMessage(0, MessageType.Get, Command.WhoIAm));
+                    //SendSingle(e.NodeId, new BinarySmarthomeMessage(0, MessageType.Get, Command.WhoIAm));
                     if (dt == default)
-                        WhoIAmSendTime.TryAdd(e.NodeId, (DateTime.Now.Subtract(WaitBeforeWhoIAmSendAgain), 0));
+                        _ = WhoIAmSendTime.TryAdd(e.NodeId, (DateTime.Now.Subtract(WaitBeforeWhoIAmSendAgain), 0));
                     else
                         WhoIAmSendTime[e.NodeId] = (DateTime.Now.Subtract(WaitBeforeWhoIAmSendAgain), 0);
                 }
                 if (!queuedMessages.TryGetValue(e.NodeId, out var queue))
                 {
-                    queue = new Queue<GeneralSmarthomeMessage>();
+                    queue = new Queue<BinarySmarthomeMessage>();
                     queuedMessages.Add(e.NodeId, queue);
                 }
                 queue.Enqueue(e);
@@ -107,20 +168,20 @@ namespace AppBokerASP
             if (known.MissedConnections > 0)
             {
                 known.MissedConnections = 0;
-                ConnectionReastablished?.Invoke(this, known.Id);
+                ConnectionReastablished?.Invoke(this, (known.Id, e.Parameters));
             }
 
             if (queuedMessages.TryGetValue(e.NodeId, out var messages))
             {
                 while (messages.TryDequeue(out var message))
                     MessageTypeSwitch(message);
-                queuedMessages.Remove(e.NodeId);
+                _ = queuedMessages.Remove(e.NodeId);
             }
 
             MessageTypeSwitch(e);
         }
 
-        private void MessageTypeSwitch(GeneralSmarthomeMessage e)
+        private void MessageTypeSwitch(BinarySmarthomeMessage e)
         {
             switch (e.MessageType)
             {
@@ -138,8 +199,6 @@ namespace AppBokerASP
             }
         }
 
-        internal void UpdateTime()
-        => SendBroadcast(new GeneralSmarthomeMessage(0, MessageType.Update, Command.Time, $"{{\"Date\":\"{DateTime.Now.ToString("dd.MM.yyyy HH:mm:ss")}\"}}".ToJToken()));
 
         private bool TryParseSubsRecursive(Sub sub, out List<Sub> subs)
         {
@@ -149,17 +208,17 @@ namespace AppBokerASP
                 subs.Add(sub);
                 if (!knownNodeIds.Any(x => x.Id == sub.NodeId))
                 {
-                    if (!WhoIAmSendTime.TryGetValue(sub.NodeId, out var dt) || dt.time.Add(WaitBeforeWhoIAmSendAgain) > DateTime.Now)
+                    if (!WhoIAmSendTime.TryGetValue(sub.NodeId, out var dt) /*|| dt.time.Add(WaitBeforeWhoIAmSendAgain*2) > DateTime.Now*/)
                     {
-                        //SendSingle(sub.NodeId, new GeneralSmarthomeMessage(0, MessageType.Get, Command.WhoIAm));
+                        //SendSingle(sub.NodeId, new BinarySmarthomeMessage(0, MessageType.Get, Command.WhoIAm));
                         if (dt == default)
-                            WhoIAmSendTime.TryAdd(sub.NodeId, (DateTime.Now, 0));
+                            _ = WhoIAmSendTime.TryAdd(sub.NodeId, (DateTime.Now, 0));
                         else
                             WhoIAmSendTime[sub.NodeId] = (DateTime.Now, 0);
                     }
                 }
 
-                if (sub.Subs != null && sub.Subs != default)
+                if (sub.Subs is not null and not null)
                     foreach (var item in sub.Subs)
                     {
                         if (TryParseSubsRecursive(item.Value, out var rsubs))
@@ -175,7 +234,7 @@ namespace AppBokerASP
             return true;
         }
 
-        private void HandleUpdates(GeneralSmarthomeMessage e)
+        private void HandleUpdates(BinarySmarthomeMessage e)
         {
             switch (e.Command)
             {
@@ -192,15 +251,17 @@ namespace AppBokerASP
             }
         }
 
-        private void RefreshMesh(GeneralSmarthomeMessage e)
+        private void RefreshMesh(BinarySmarthomeMessage e)
         {
             try
             {
-                var sub = e.Parameters[0].ToObject<Sub>();
-                if (!TryParseSubsRecursive(sub, out var rsubs))
+                var str = Encoding.UTF8.GetString(e.Parameters[0]);
+                logger.Debug(str);
+                var sub = JsonConvert.DeserializeObject<Sub>(str);
+                if (sub is null || !TryParseSubsRecursive(sub, out var rsubs))
                     return;
                 var subs = rsubs.Distinct().ToDictionary(x => x.NodeId, x => x);
-                var lostSubs = new List<long>();
+                var lostSubs = new List<uint>();
 
                 foreach (var item in knownNodeIds) // Check our subs with incoming list
                 {
@@ -212,7 +273,7 @@ namespace AppBokerASP
                         {
                             item.MissedConnections = 0;
                             if (Program.DeviceManager.Devices.TryGetValue(item.Id, out var dev))
-                                dev.Reconnect();
+                                SendSingle((uint)dev.Id, new BinarySmarthomeMessage(0, MessageType.Get, Command.WhoIAm));
                             else
                                 lostSubs.Add(item.Id);
                         }
@@ -222,8 +283,8 @@ namespace AppBokerASP
                 {
                     if (lostSubs.Contains(item.Key))
                     {
-                        if (subs.ContainsKey(item.Key))
-                            lostSubs.Remove(item.Key);
+                        if (subs.ContainsKey((uint)item.Key))
+                            _ = lostSubs.Remove(item.Key);
                     }
                 }
                 foreach (var id in lostSubs)
@@ -231,8 +292,8 @@ namespace AppBokerASP
                     var knownId = knownNodeIds.FirstOrDefault(x => x.Id == id);
                     ConnectionLost?.Invoke(this, id);
                     if (knownId != default)
-                        if (knownId.MissedConnections > 4)
-                            knownNodeIds.Remove(knownId);
+                        if (knownId.MissedConnections > 15)
+                            _ = knownNodeIds.Remove(knownId);
                         else
                             knownId.MissedConnections++;
                 }
@@ -240,66 +301,141 @@ namespace AppBokerASP
             }
             catch (Exception ex)
             {
-                logger.Error(nameof(RefreshMesh) + ex);
+                logger.Error(ex);
             }
         }
 
-        private void HandleGets(GeneralSmarthomeMessage e)
+        private void HandleGets(BinarySmarthomeMessage e)
         {
             if (e.Command == Command.Time)
-                SendBroadcast(new GeneralSmarthomeMessage(e.NodeId, MessageType.Update, Command.Time, $"{{\"Date\":\"{DateTime.Now.ToString("dd.MM.yyyy HH:mm:ss")}\"}}".ToJToken()));
+                SendTimeUpdate(null);
             else
                 SingleGetMessageReceived?.Invoke(this, e);
         }
 
-        private void HandleOptions(GeneralSmarthomeMessage e) => SingleOptionsMessageReceived?.Invoke(this, e);
+        private void HandleOptions(BinarySmarthomeMessage e) => SingleOptionsMessageReceived?.Invoke(this, e);
 
-        public void SendSingle<T>(long destination, T message)
+        public void SendSingle<T>(uint destination, T message) where T : BinarySmarthomeMessage
         {
             try
             {
-                serverSocket.SendToAllClients(PackageType.SINGLE, message.ToJson(), destination);
+                using var ms = new MemoryStream();
+                message.Serialize(ms);
+                logger.Debug($"{PackageType.SINGLE}: NodeId: {destination}, Command: {message.Command}, MessageType: {message.MessageType}, ParamsAmount: {message.Parameters.Count}, " + string.Join(", ", message.Parameters.Select(x => BitConverter.ToString(x))));
+                serverSocket.SendToAllClients(PackageType.SINGLE, ms.ToArray(), destination);
             }
             catch (Exception e)
             {
                 Console.WriteLine(e);
+                logger.Error(e);
             }
         }
-
-        public void SendBroadcast<T>(T message)
+        public void SendSingle(uint destination, Memory<byte> message)
         {
             try
             {
-                serverSocket.SendToAllClients(PackageType.BROADCAST, message.ToJson());
+                serverSocket.SendToAllClients(PackageType.SINGLE, message, destination, false);
             }
             catch (Exception e)
             {
                 Console.WriteLine(e);
+                logger.Error(e);
             }
         }
+        //public void SendCustomSingle<T>(long destination, PackageType type, T message)
+        //{
+        //    try
+        //    {
+        //        serverSocket.SendToAllClients(type, message.ToJson(), destination);
+        //    }
+        //    catch (Exception e)
+        //    {
+        //        Console.WriteLine(e);
+        //        logger.Error(e);
+        //    }
+        //}
 
-        public void SendToBridge<T>(T message)
+        public void SendBroadcast<T>(T message) where T : BinarySmarthomeMessage
         {
             try
             {
-                serverSocket.SendToAllClients(PackageType.BRIDGE, message.ToJson(), nodeID);
+                using var ms = new MemoryStream();
+                message.Serialize(ms);
+                logger.Debug($"{PackageType.BROADCAST}: Command: {message.Command}, MessageType: {message.MessageType}, ParamsAmount: {message.Parameters.Count}, " + string.Join(", ", message.Parameters.Select(x => BitConverter.ToString(x))));
+                serverSocket.SendToAllClients(PackageType.BROADCAST, ms.ToArray());
             }
             catch (Exception e)
             {
                 Console.WriteLine(e);
+                logger.Error(e);
             }
         }
 
-        private void WhoAmITask(object o)
+        public void SendBroadcast(Memory<byte> message)
         {
-            var toDelete = new Dictionary<long, (DateTime time, int count)>();
+            try
+            {
+                serverSocket.SendToAllClients(PackageType.BROADCAST, message, false);
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+                logger.Error(e);
+            }
+        }
+        //public void SendCustomBroadcast<T>(PackageType type, T message)
+        //{
+        //    try
+        //    {
+        //        serverSocket.SendToAllClients(type, message.ToJson(), 0);
+        //    }
+        //    catch (Exception e)
+        //    {
+        //        Console.WriteLine(e);
+        //        logger.Error(e);
+        //    }
+        //}
+
+        public void SendToBridge<T>(T message) where T : BinarySmarthomeMessage
+        {
+            try
+            {
+                using var ms = new MemoryStream();
+                message.Serialize(ms);
+                logger.Debug($"{PackageType.BRIDGE}: Command: {message.Command}, MessageType: {message.MessageType}, ParamsAmount: {message.Parameters.Count}, " + string.Join(", ", message.Parameters.Select(x => BitConverter.ToString(x))));
+
+                serverSocket.SendToAllClients(PackageType.BRIDGE, ms.ToArray(), nodeID);
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+                logger.Error(e);
+            }
+        }
+
+        private void TryCatchWhoAmITask(object? o)
+        {
+            try
+            {
+                WhoAmITask(o);
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex);
+                logger.Error(ex);
+            }
+        }
+
+        private void WhoAmITask(object? o)
+        {
+            var toDelete = new Dictionary<uint, (DateTime time, int count)>();
             try
             {
                 foreach (var item in WhoIAmSendTime)
                 {
                     if (item.Value.time.Add(WaitBeforeWhoIAmSendAgain) < DateTime.Now)
                     {
-                        SendSingle(item.Key, new GeneralSmarthomeMessage(0, MessageType.Get, Command.WhoIAm));
+                        SendSingle(item.Key, new BinarySmarthomeMessage(0, MessageType.Get, Command.WhoIAm));
                         WhoIAmSendTime[item.Key] = (DateTime.Now, item.Value.count + 1);
                     }
                     if (item.Value.count > 20)
@@ -308,16 +444,24 @@ namespace AppBokerASP
 
                 foreach (var item in toDelete)
                 {
-                    WhoIAmSendTime.TryRemove(item.Key, out var val);
+                    _ = WhoIAmSendTime.TryRemove(item.Key, out var val);
                 }
                 toDelete.Clear();
             }
             catch (Exception e)
             {
                 Console.WriteLine(e);
+                logger.Error(e);
             }
         }
 
+        public void Dispose()
+        {
+            foreach (var item in timers)
+            {
+                item.Dispose();
+            }
+        }
     }
 }
 
