@@ -2,7 +2,10 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using System.Reflection;
 using System.Text.Json;
+using AppBokerASP.Devices.Painless;
+
 using Microsoft.AspNetCore.SignalR;
 using Newtonsoft.Json.Linq;
 using PainlessMesh;
@@ -11,7 +14,7 @@ namespace AppBokerASP.Devices
 {
     public abstract class Device
     {
-        public event EventHandler<Device> PrintableInfoChanged;
+        public event EventHandler<Device>? PrintableInfoChanged;
 
         public long Id { get; set; }
         //public List<string> PrintableInformation { get; set; } = new List<string>();
@@ -20,6 +23,7 @@ namespace AppBokerASP.Devices
         public bool ShowInApp { get; set; }
         public string FriendlyName { get; set; }
         public bool IsConnected { get; set; }
+
         protected readonly NLog.Logger logger;
 
         public Device(long nodeId)
@@ -28,18 +32,19 @@ namespace AppBokerASP.Devices
             TypeName = GetType().Name;
             IsConnected = true;
             logger = NLog.LogManager.GetCurrentClassLogger();
+            FriendlyName = "";
         }
 
         public virtual void UpdateFromApp(Command command, List<JToken> parameters) { }
         public virtual void OptionsFromApp(Command command, List<JToken> parameters) { }
 
-        public virtual dynamic GetConfig() { return null; }
+        public virtual dynamic? GetConfig() { return null; }
 
         public virtual async void SendLastData(IClientProxy client) => await client.SendAsync("Update", this);
         public virtual void SendLastData(List<IClientProxy> clients) => clients.ForEach(async x => await x.SendAsync("Update", this));
         public virtual void SendDataToAllSubscribers() => Subscribers.ForEach(x => SendLastData(x.ClientProxy));
 
         public virtual void StopDevice() => IsConnected = false;
-        public virtual void Reconnect() => IsConnected = true;
+        public virtual void Reconnect(ByteLengthList parameter) => IsConnected = true;
     }
 }
