@@ -16,8 +16,8 @@ namespace AppBokerASP.Devices.Zigbee
         public DateTime LastReceived { get; set; }
         [JsonPropertyName("linkQuality")]
         public byte Link_Quality { get; set; }
-        public bool Available { get; set; }
-        public string AdapterWithId { get; set; }
+        public bool Available { get; set; } 
+        public string AdapterWithId { get; set; } = "";
 
         private readonly ReadOnlyCollection<PropertyInfo> propertyInfos;
 
@@ -34,7 +34,7 @@ namespace AppBokerASP.Devices.Zigbee
             var prop = propertyInfos.FirstOrDefault(x => x.Name.ToLower() == ioBrokerObject.ValueName);
             if (prop == default || ioBrokerObject.ValueParameter.Value == null)
                 return;
-            if ((prop.PropertyType == typeof(float) || prop.PropertyType == typeof(double)) && ioBrokerObject.ValueParameter.Value.ToObject<string>().Contains(","))
+            if ((prop.PropertyType == typeof(float) || prop.PropertyType == typeof(double)) && ioBrokerObject.ValueParameter.Value.ToObject<string>()!.Contains(","))
             {
                 var strValue = ioBrokerObject.ValueParameter.Value.ToObject<string>();
                 if (double.TryParse(strValue, out var val))
@@ -49,7 +49,7 @@ namespace AppBokerASP.Devices.Zigbee
         public virtual List<IoBrokerHistory> ReadHistoryJSON(DateTime date)
         {
 #if DEBUG
-            var files = Directory.GetFiles($"F:\\Temp\\Neuer Ordner\\{date.ToString("yyyyMMdd")}\\", $"history.{AdapterWithId}*");
+            var files = Directory.GetFiles($"F:\\Temp\\Neuer Ordner\\{date:yyyyMMdd}\\", $"history.{AdapterWithId}*");
 #else
             var files = Directory.GetFiles($"/home/pi/IoBrokerHistory/{date.ToString("yyyyMMdd")}/", $"history.{AdapterWithId}*");
 #endif
@@ -57,9 +57,9 @@ namespace AppBokerASP.Devices.Zigbee
             var list = new List<IoBrokerHistory>();
             foreach (var file in files)
             {
-                var hist = Newtonsoft.Json.JsonConvert.DeserializeObject<HistoryRecord[]>(File.ReadAllText(file));
+                var hist = Newtonsoft.Json.JsonConvert.DeserializeObject<HistoryRecord[]>(File.ReadAllText(file))!;
 
-                list.Add(new IoBrokerHistory(hist, file.Split('.').SkipLast(1).LastOrDefault()));
+                list.Add(new IoBrokerHistory(hist, file!.Split('.').SkipLast(1).Last()));
             }
             return list;
             //return Newtonsoft.Json.JsonConvert.DeserializeObject<IoBrokerHistory>(File.ReadAllText($"/home/pi/ioBrokerHistory/{date.ToString("yyyyMMdd")}/history.{AdapterWithId}.{property}.json"));
@@ -67,7 +67,7 @@ namespace AppBokerASP.Devices.Zigbee
 
         public virtual IoBrokerHistory ReadHistoryJSON(DateTime date, string property)
         {
-            return Newtonsoft.Json.JsonConvert.DeserializeObject<IoBrokerHistory>(File.ReadAllText($"/home/pi/ioBrokerHistory/{date.ToString("yyyyMMdd")}/history.{AdapterWithId}.{property}.json"));
+            return Newtonsoft.Json.JsonConvert.DeserializeObject<IoBrokerHistory>(File.ReadAllText($"/home/pi/ioBrokerHistory/{date:yyyyMMdd}/history.{AdapterWithId}.{property}.json"))!;
         }
 
 
