@@ -1,5 +1,7 @@
 ﻿using Newtonsoft.Json.Linq;
+
 using PainlessMesh;
+
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -28,7 +30,7 @@ namespace AppBokerASP.Devices.Zigbee
             {
                 case Command.Delay:
                     Transition_Time = parameters[0].ToObject<float>();
-                    UpdateZigbeeDeviceRequest(nameof(Transition_Time).ToLower(), Transition_Time);
+                    _ = UpdateZigbeeDeviceRequest(nameof(Transition_Time).ToLower(), Transition_Time);
                     break;
             }
         }
@@ -39,15 +41,19 @@ namespace AppBokerASP.Devices.Zigbee
             {
                 case Command.Temp:
                     ColorTemp = parameters[0].ToObject<int>();
-                    UpdateZigbeeDeviceRequest(nameof(ColorTemp).ToLower(), ColorTemp);
+                    _ = UpdateZigbeeDeviceRequest(nameof(ColorTemp).ToLower(), ColorTemp);
                     break;
                 case Command.Brightness:
                     Brightness = parameters[0].ToObject<byte>();
-                    UpdateZigbeeDeviceRequest(nameof(Brightness).ToLower(), Brightness);
+                    _ = UpdateZigbeeDeviceRequest(nameof(Brightness).ToLower(), Brightness);
+                    break;
+                case Command.SingleColor:
+                    State = true;
+                    _ = UpdateZigbeeDeviceRequest(nameof(State).ToLower(), State.ToString().ToLower());
                     break;
                 case Command.Off:
-                    State = !State;
-                    UpdateZigbeeDeviceRequest(nameof(State).ToLower(), State.ToString().ToLower());
+                    State = false;
+                    _ = UpdateZigbeeDeviceRequest(nameof(State).ToLower(), State.ToString().ToLower());
                     break;
                 default:
                     break;
