@@ -1,12 +1,29 @@
-﻿using System;
+﻿using Newtonsoft.Json.Linq;
+
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
 namespace AppBokerASP.IOBroker
 {
+
     public class IoBrokerStateResponse
     {
+        public IoBrokerStateResponse(JToken val, bool ack, long ts, int q, string from, long lc, string type, Common common, string id, Acl acl)
+        {
+            this.val = val;
+            this.ack = ack;
+            this.ts = ts;
+            this.q = q;
+            this.from = from;
+            this.lc = lc;
+            this.type = type;
+            this.common = common;
+            _id = id;
+            this.acl = acl;
+        }
+
         public Newtonsoft.Json.Linq.JToken val { get; set; }
         public bool ack { get; set; }
         public long ts { get; set; }
