@@ -1,5 +1,6 @@
 ﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -13,33 +14,39 @@ namespace AppBokerASP.IOBroker
     {
         public long Id { get; set; }
 
-        public static bool TryParse(string s, out IoBrokerZigbee obj)
+        public IoBrokerZigbee(long id, BrokerEvent eventName, string adapterName, byte adapterInstance, string valueName, Parameter valueParameter)
+            : base(eventName, adapterName, adapterInstance, valueName, valueParameter)
         {
-            obj = new IoBrokerZigbee();
-            var i = s[2..].IndexOf('"')+2;
+            Id = id;
+        }
+
+        public static bool TryParse(string s, out IoBrokerZigbee? obj)
+        {
+            obj = null;
+            var i = s[2..].IndexOf('"') + 2;
             if (!Enum.TryParse<BrokerEvent>(s[2..i], true, out var en))
                 return false;
-            obj.EventName = en;
 
-            i = s[++i..].IndexOf('"') + 1+i;
-            var split = s[i..(s[i..].IndexOf('"')+i)].Split('.');
+            i = s[++i..].IndexOf('"') + 1 + i;
+            var split = s[i..(s[i..].IndexOf('"') + i)].Split('.');
 
             int splitIndex = -1;
             byte instanceNum;
-            while (!byte.TryParse(split[++splitIndex], out instanceNum) && splitIndex+1 < split.Length)
+            while (!byte.TryParse(split[++splitIndex], out instanceNum) && splitIndex + 1 < split.Length)
             { }
-            if (splitIndex+1 == split.Length)
+            if (splitIndex + 1 == split.Length)
                 return false;
-            obj.AdapterInstance = instanceNum;
-            obj.AdapterName = string.Join('.', split[..(splitIndex)]);
+            var adapterInstance = instanceNum;
+            var adapterName = string.Join('.', split[..(splitIndex)]);
 
             if (!long.TryParse(split[splitIndex + 1], System.Globalization.NumberStyles.HexNumber, null, out var id))
                 return false;
-            obj.Id = id;
-            obj.ValueName = split.Last();
+            var valueName = split.Last();
 
-            obj.ValueParameter = s[(s[i..].IndexOf(',')+i+1)..^1].ToDeObject<Parameter>();
-            
+            var valueParameter = s[(s[i..].IndexOf(',') + i + 1)..^1].ToDeObject<Parameter>();
+
+            obj = new IoBrokerZigbee(id, en, adapterName, adapterInstance, valueName, valueParameter);
+
             return true;
         }
 
@@ -48,19 +55,27 @@ namespace AppBokerASP.IOBroker
 
     public class IoBrokerObject
     {
+        public IoBrokerObject(BrokerEvent eventName, string adapterName, byte adapterInstance, string valueName, Parameter valueParameter)
+        {
+            EventName = eventName;
+            AdapterName = adapterName;
+            AdapterInstance = adapterInstance;
+            ValueName = valueName;
+            ValueParameter = valueParameter;
+        }
+
         public BrokerEvent EventName { get; set; }
         public string AdapterName { get; set; }
         public byte AdapterInstance { get; set; }
         public string ValueName { get; set; }
         public Parameter ValueParameter { get; set; }
 
-        public static bool TryParse(string s, out IoBrokerObject obj)
+        public static bool TryParse(string s, out IoBrokerObject? obj)
         {
-            obj = new IoBrokerObject();
+            obj = null;
             var i = s[2..].IndexOf('"');
             if (!Enum.TryParse<BrokerEvent>(s[2..i], out var en))
                 return false;
-            obj.EventName = en;
 
             i = s[++i..].IndexOf('"') + 1;
             var split = s[i..s[i..].IndexOf('"')].Split('.');
@@ -71,13 +86,13 @@ namespace AppBokerASP.IOBroker
             { }
             if (splitIndex < split.Length)
                 return false;
-            obj.AdapterInstance = instanceNum;
-            obj.AdapterName = string.Join('.', split[..(splitIndex - 1)]);
+            var adapterName = string.Join('.', split[..(splitIndex - 1)]);
 
-            obj.ValueName = split.Last();
+            var valueName = split.Last();
 
-            obj.ValueParameter = s[s[i..].IndexOf(',')..].ToDeObject<Parameter>();
+            var valueParameter = s[s[i..].IndexOf(',')..].ToDeObject<Parameter>();
 
+            obj = new IoBrokerObject(en, adapterName, instanceNum, valueName, valueParameter);
             return true;
         }
     }
@@ -86,6 +101,12 @@ namespace AppBokerASP.IOBroker
 
     public class Parameter
     {
+        public Parameter(JToken value)
+        {
+            Value = value;
+            From = "";
+        }
+
         [JsonProperty("val")]
         public JToken Value { get; set; }
         [JsonProperty("ack")]
