@@ -9,8 +9,14 @@
         NODE_SYNC_REPLY = 6,
         BROADCAST = 8,  //application data for everyone
         SINGLE = 9,   //application data for a single node
-        BRIDGE = 255, 
+        OTA_ANNOUNCE = 10,
+        OTA_REQUEST = 11,
+        OTA_REPLY = 12,
+        BRIDGE = 255,
+         
     };
+
+    
 
     public enum MessageType
     {
@@ -43,7 +49,9 @@
         Reverse,
         SingleColor,
         DeviceMapping,
-        Calibration
+        Calibration,
+        Ota,
+        OtaPart
 
     };
 }
